@@ -16,10 +16,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 # In Windows, build-name is used as the major, minor, and patch parts
 # of the product and file versions while build-number is used as the build suffix.
-version: 1.6.0+13
+version: 2.4.2+30
 
 environment:
-  sdk: '>=2.19.3 <3.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -33,36 +33,42 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.8
   delayed_display: ^2.0.0
-  shared_preferences: ^2.0.18
-  sqflite: ^2.2.5
-  path_provider: ^2.0.13
-  path: ^1.8.2
-  chat_gpt_sdk: ^2.0.6
-  http: ^0.13.5
-  cached_network_image: ^3.2.3
-  loading_animation_widget: ^1.2.0+4
-  flutter_launcher_icons: ^0.13.1
-  share_plus: ^7.0.2
-  url_launcher: ^6.1.11
-  fluttertoast: ^8.2.2
+  shared_preferences: ^2.5.3
+  sqflite: ^2.4.2
+  path_provider: ^2.1.5
+  path: ^1.9.1
+  crypto: ^3.0.3
+  openai_dart: ^0.6.0+1
+  http: ^1.5.0
+  cached_network_image: ^3.4.1
+  loading_animation_widget: ^1.3.0
+  flutter_launcher_icons: ^0.14.4
+  share_plus: ^12.0.1
+  url_launcher: ^6.3.2
+  fluttertoast: ^9.0.0
   sliding_up_panel: ^2.0.0+1
-  oktoast: ^3.3.1
-  tutorial_coach_mark: ^1.2.9
-  easy_localization: ^3.0.2
-  toggle_switch: ^2.1.0
+  oktoast: ^3.4.0
+  easy_localization: ^3.0.8
+  toggle_switch: ^2.3.0
   auto_size_text: ^3.0.0
-  firebase_analytics: ^10.6.3
-  firebase_core: ^2.21.0
-  in_app_review: ^2.0.8
-  google_mobile_ads: ^3.1.0
+  firebase_analytics: ^12.0.3
+  firebase_core: ^4.2.0
+  in_app_review: ^2.0.11
+  google_mobile_ads: ^5.3.1
+  cloud_firestore: ^6.0.3
+  firebase_messaging: ^16.0.4
+  flutter_local_notifications: ^19.5.0
+  timezone: ^0.10.0
+  file_picker: ^10.3.7
+  csv: ^6.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^6.0.0
 
 flutter_launcher_icons:
   android: true
