@@ -1,55 +1,67 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:ui';
-import 'package:chat_gpt_sdk/chat_gpt_sdk.dart';
-import 'package:delayed_display/delayed_display.dart';
+import 'package:openai_dart/openai_dart.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:oktoast/oktoast.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'package:toggle_switch/toggle_switch.dart';
-import 'package:tutorial_coach_mark/tutorial_coach_mark.dart';
-import 'package:watch_next/pages/recommandation_results_page.dart';
-import 'package:watch_next/pages/settings_page.dart';
+import 'package:watch_next/pages/recommendation_loading_page.dart';
+import 'package:watch_next/services/feedback_service.dart';
+import 'package:watch_next/services/notification_service.dart';
 import 'package:watch_next/utils/secrets.dart';
-import 'package:watch_next/widgets/toast_widget.dart';
+import 'package:watch_next/utils/prompts.dart';
+import 'package:watch_next/widgets/feedback_dialog.dart';
+import 'package:watch_next/widgets/main_menu/examples_dialog.dart';
+import 'package:watch_next/widgets/main_menu/main_menu_top_bar.dart';
+import 'package:watch_next/widgets/main_menu/media_type_switch.dart';
+import 'package:watch_next/widgets/main_menu/prompt_input_widget.dart';
+import 'package:watch_next/widgets/main_menu/query_settings_panel.dart';
+import 'package:watch_next/widgets/shared/toast_widget.dart';
 
 class MainMenuPage extends StatefulWidget {
-  const MainMenuPage({Key? key}) : super(key: key);
+  const MainMenuPage({super.key});
 
   @override
   State<MainMenuPage> createState() => _MainMenuPageState();
 }
 
 class _MainMenuPageState extends State<MainMenuPage> {
-  int currentIndex = -1;
-
-  final openAI = OpenAI.instance
-      .build(token: openApiKey, baseOption: HttpSetup(receiveTimeout: const Duration(seconds: 20)), enableLog: true);
-
+  late final OpenAIClient openAI;
   final _controller = TextEditingController();
+  final GlobalKey textFieldKey = GlobalKey();
 
   bool isLongEnough = false;
   bool isValidQuery = false;
   bool enableLoading = false;
-
-  GlobalKey textFieldKey = GlobalKey();
-  GlobalKey goButtonKey = GlobalKey();
-
-  late TutorialCoachMark tutorialCoachMark;
   bool noInternet = false;
-  int typeIsMovie = 0; //0 = movie , 1 = show
+  int typeIsMovie = 0;
+  QuerySettings _querySettings = const QuerySettings();
 
   @override
   void initState() {
-    createTutorial();
     super.initState();
-    _controller.addListener(checkLength);
-    _controller.text = ' ';
-    Timer(const Duration(seconds: 2), () {
-      showTutorial();
+    openAI = OpenAIClient(apiKey: openApiKey);
+    _controller.addListener(_checkLength);
+    _loadQuerySettings();
+    // Reschedule notification with proper translations once context is available
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _rescheduleNotificationWithContext();
     });
+  }
+
+  Future<void> _rescheduleNotificationWithContext() async {
+    if (mounted) {
+      await NotificationService.rescheduleWithTranslations(context);
+    }
+  }
+
+  Future<void> _loadQuerySettings() async {
+    final settings = await QuerySettingsService.load();
+    if (mounted) {
+      setState(() => _querySettings = settings);
+    }
   }
 
   @override
@@ -58,406 +70,157 @@ class _MainMenuPageState extends State<MainMenuPage> {
     super.dispose();
   }
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color.fromRGBO(11, 14, 23, 1),
-      body: body(),
-    );
-  }
-
-  Widget body() {
-    return SizedBox(
-      height: MediaQuery.of(context).size.height,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16),
-        child: Column(
-          children: [
-            const SizedBox(height: 32),
-            topBar(),
-            Expanded(
-              child: Container(),
-            ),
-            description(),
-            const SizedBox(height: 32),
-            switchWidget(),
-            Expanded(
-              child: Container(),
-            ),
-            examplesWidget(),
-            const SizedBox(
-              height: 16,
-            ),
-            Align(
-              alignment: Alignment.topLeft,
-              child: promptInput(),
-            ),
-            const SizedBox(height: 32),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget topBar() {
-    return Row(
-      children: [
-        const SizedBox(
-          width: 48,
-        ),
-        Expanded(
-          child: Container(),
-        ),
-        DelayedDisplay(
-          fadingDuration: const Duration(milliseconds: 1000),
-          child: Text(
-            "hey_there".tr(),
-            style: Theme.of(context).textTheme.bodyMedium,
-          ),
-        ),
-        Expanded(
-          child: Container(),
-        ),
-        settingsButton(),
-      ],
-    );
-  }
-
-  Widget settingsButton() {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.grey[400],
-        borderRadius: BorderRadius.circular(25),
-      ),
-      child: IconButton(
-        icon: Icon(
-          Icons.settings,
-          color: Colors.grey[900],
-          size: 28,
-        ),
-        onPressed: () {
-          FirebaseAnalytics.instance.logEvent(
-            name: 'opened_settings',
-          );
-          Navigator.push(
-            context,
-            MaterialPageRoute(
-              builder: (context) => const SettingsPage(),
-            ),
-          );
-        },
-      ),
-    );
-  }
-
-  Widget examplesWidget() {
-    return Align(
-      alignment: Alignment.centerRight,
-      child: Container(
-        decoration: BoxDecoration(
-          color: Colors.grey[400],
-          borderRadius: BorderRadius.circular(25),
-        ),
-        child: IconButton(
-          icon: Icon(
-            Icons.help_outline_rounded,
-            color: Colors.grey[900],
-            size: 26,
-          ),
-          onPressed: () {
-            FirebaseAnalytics.instance.logEvent(
-              name: 'opened_examples',
-              parameters: <String, dynamic>{
-                "type": typeIsMovie == 0 ? "movie" : "show",
-              },
-            );
-            typeIsMovie == 0 ? showExamples() : showExamplesShows();
-          },
-        ),
-      ),
-    );
-  }
-
-  Widget switchWidget() {
-    return ToggleSwitch(
-      minWidth: 110.0,
-      initialLabelIndex: typeIsMovie,
-      cornerRadius: 15.0,
-      animate: true,
-      animationDuration: 400,
-      activeFgColor: Colors.white,
-      inactiveBgColor: Colors.grey[900],
-      inactiveFgColor: Colors.white,
-      totalSwitches: 2,
-      labels: ['movie'.tr(), 'tv_show'.tr()],
-      activeBgColors: const [
-        [Colors.orange],
-        [Colors.orange],
-      ],
-      onToggle: (index) {
-        setState(() {
-          typeIsMovie = index!;
-        });
-      },
-    );
-  }
-
-  Widget description() {
-    return Text(
-      "find_something".tr(),
-      textAlign: TextAlign.center,
-      style: Theme.of(context).textTheme.displayMedium,
-    );
-  }
-
-  Widget promptInput() {
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        SizedBox(
-          width: MediaQuery.of(context).size.width - 88,
-          child: TextField(
-            key: textFieldKey,
-            autofocus: false,
-            showCursor: true,
-            maxLength: 80,
-            maxLines: 3,
-            minLines: 1,
-            controller: _controller,
-            cursorColor: Colors.orange,
-            style: Theme.of(context).textTheme.titleMedium,
-            decoration: InputDecoration(
-              filled: true,
-              fillColor: const Color.fromARGB(255, 44, 46, 56),
-              helperText: "complete_sentence".tr(),
-              prefixText: typeIsMovie == 0 ? "recommend_a_movie".tr() : "recommend_a_show".tr(),
-              prefixStyle: Theme.of(context).textTheme.displaySmall!.copyWith(fontSize: 12, letterSpacing: 0.5),
-              helperStyle: TextStyle(
-                color: Colors.grey[500],
-                fontSize: 12,
-              ),
-              contentPadding: const EdgeInsets.only(left: 14.0, bottom: 10.0, top: 10.0),
-              focusedBorder: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(25),
-              ),
-              enabledBorder: UnderlineInputBorder(
-                borderRadius: BorderRadius.circular(25),
-              ),
-            ),
-          ),
-        ),
-        const SizedBox(
-          width: 8,
-        ),
-        goButton(),
-      ],
-    );
-  }
-
-  Widget goButton() {
-    return Container(
-      height: 48,
-      width: 48,
-      decoration: BoxDecoration(
-        color: isLongEnough ? Colors.orange : Colors.grey[700],
-        borderRadius: BorderRadius.circular(50),
-      ),
-      child: Center(
-        child: IconButton(
-          key: goButtonKey,
-          onPressed: () async {
-            isLongEnough ? goButtonPressed() : null;
-          },
-          icon: Icon(
-            Icons.arrow_forward,
-            size: 32,
-            color: Colors.grey[900],
-          ),
-        ),
-      ),
-    );
-  }
-
-  void checkLength() {
+  void _checkLength() {
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (_controller.text.length > 5 && mounted) {
+      if (mounted) {
         setState(() {
-          isLongEnough = true;
-        });
-      }
-      if (_controller.text.length < 5 && mounted) {
-        setState(() {
-          isLongEnough = false;
+          isLongEnough = _controller.text.length >= 5;
         });
       }
     });
   }
 
-  validateQuery() async {
-    final request = ChatCompleteText(
-      messages: [
-        Messages(
-            role: Role.assistant,
-            content: typeIsMovie == 0
-                ? 'validation_prompt'.tr() + _controller.text
-                : 'validation_prompt_series'.tr() + _controller.text),
-      ],
-      maxToken: 400,
-      model: GptTurbo0301ChatModel(),
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () => FocusScope.of(context).unfocus(),
+      child: Scaffold(
+        backgroundColor: Theme.of(context).colorScheme.primary,
+        resizeToAvoidBottomInset: true,
+        body: _buildBody(),
+      ),
     );
-
-    final response = await openAI.onChatCompletion(request: request);
-    if (response!.choices[0].message!.content == "YES" && mounted) {
-      setState(() {
-        isValidQuery = true;
-        enableLoading = false;
-      });
-    } else {
-      setState(() {
-        isValidQuery = false;
-        enableLoading = false;
-      });
-    }
   }
 
-  Future<void> checkConnection() async {
+  Widget _buildBody() {
+    final bottomInset = MediaQuery.of(context).viewInsets.bottom;
+    final screenHeight = MediaQuery.of(context).size.height;
+    final topPadding = MediaQuery.of(context).padding.top;
+    final bottomPadding = MediaQuery.of(context).padding.bottom;
+    const tabBarHeight = 70;
+
+    final availableHeight = screenHeight - topPadding - bottomPadding - tabBarHeight - bottomInset;
+
+    return SafeArea(
+      bottom: false,
+      child: SizedBox(
+        height: availableHeight,
+        child: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 16),
+          child: Column(
+            children: [
+              const SizedBox(height: 8),
+              const MainMenuTopBar(),
+              const Spacer(),
+              _buildDescription(),
+              const SizedBox(height: 32),
+              MediaTypeSwitch(
+                currentIndex: typeIsMovie,
+                onToggle: (index) => setState(() => typeIsMovie = index),
+              ),
+              const Spacer(),
+              Row(
+                mainAxisAlignment: MainAxisAlignment.end,
+                children: [
+                  QuerySettingsButton(
+                    settings: _querySettings,
+                    onSettingsChanged: (settings) {
+                      setState(() => _querySettings = settings);
+                    },
+                    isMovie: typeIsMovie == 0,
+                  ),
+                  const SizedBox(width: 12),
+                  ExamplesButton(isMovie: typeIsMovie == 0),
+                ],
+              ),
+              const SizedBox(height: 16),
+              PromptInputWidget(
+                controller: _controller,
+                textFieldKey: textFieldKey,
+                isLongEnough: isLongEnough,
+                enableLoading: enableLoading,
+                onGoPressed: _onGoPressed,
+              ),
+              const SizedBox(height: 16),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildDescription() {
+    return Text(
+      "find_something".tr(),
+      textAlign: TextAlign.center,
+      style: Theme.of(context).textTheme.displayMedium?.copyWith(
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+            height: 1.4,
+          ),
+    );
+  }
+
+  Future<void> _checkConnection() async {
     try {
       final result = await InternetAddress.lookup('example.com');
       if (result.isNotEmpty && result[0].rawAddress.isNotEmpty) {
-        setState(() {
-          noInternet = false;
-        });
+        setState(() => noInternet = false);
       }
     } on SocketException catch (_) {
-      setState(() {
-        noInternet = true;
-      });
+      setState(() => noInternet = true);
     }
   }
 
-  void showTutorial() async {
-    SharedPreferences prefs = await SharedPreferences.getInstance();
-    int showed = prefs.getInt('showed_tutorial') ?? 0;
-    if (showed == 0 && mounted) {
-      tutorialCoachMark.show(context: context);
-    }
-    prefs.setInt("showed_tutorial", 1);
-  }
+  Future<void> _validateQuery() async {
+    try {
+      final response = await openAI.createChatCompletion(
+        request: CreateChatCompletionRequest(
+          model: ChatCompletionModel.modelId('gpt-5-mini'),
+          messages: [
+            ChatCompletionMessage.system(
+              content: typeIsMovie == 0 ? validationPromptMovie : validationPromptSeries,
+            ),
+            ChatCompletionMessage.user(
+              content: ChatCompletionUserMessageContent.string(_controller.text),
+            ),
+          ],
+          reasoningEffort: ReasoningEffort.low,
+        ),
+      );
 
-  void createTutorial() {
-    tutorialCoachMark = TutorialCoachMark(
-      targets: _createTargets(),
-      colorShadow: Colors.grey[900]!,
-      textSkip: "close".tr(),
-      paddingFocus: 10,
-      opacityShadow: 0.5,
-      focusAnimationDuration: const Duration(seconds: 2),
-      imageFilter: ImageFilter.blur(sigmaX: 8, sigmaY: 8),
-      onClickTarget: (target) {
-        if (target.keyTarget == textFieldKey) {
-          setState(() {
-            _controller.text = "with_action".tr();
-          });
-        } else {
-          goButtonPressed();
-        }
-      },
-      onClickOverlay: (target) {
-        setState(
-          () {
-            _controller.text = "with_action".tr();
-          },
+      if (mounted) {
+        setState(() {
+          isValidQuery = response.choices.first.message.content == "YES";
+          enableLoading = false;
+        });
+      }
+    } catch (e) {
+      FirebaseAnalytics.instance.logEvent(
+        name: 'api_error',
+        parameters: <String, Object>{
+          'error': 'validation_query_failed',
+          'message': e.toString(),
+        },
+      );
+
+      if (mounted) {
+        setState(() => enableLoading = false);
+        showToastWidget(
+          ToastWidget(
+            title: "error_occurred".tr(),
+            icon: const Icon(Icons.error_outline, color: Colors.red, size: 36),
+          ),
+          duration: const Duration(seconds: 4),
         );
-      },
-    );
+      }
+    }
   }
 
-  List<TargetFocus> _createTargets() {
-    List<TargetFocus> targets = [];
-    targets.add(
-      TargetFocus(
-        shape: ShapeLightFocus.RRect,
-        identify: "textFieldKey",
-        keyTarget: textFieldKey,
-        alignSkip: Alignment.topRight,
-        enableOverlayTab: true,
-        contents: [
-          TargetContent(
-            align: ContentAlign.top,
-            builder: (context, controller) {
-              return Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text(
-                    "step_one".tr(),
-                    style: Theme.of(context).textTheme.displayMedium,
-                  ),
-                ],
-              );
-            },
-          ),
-          TargetContent(
-            align: ContentAlign.bottom,
-            builder: (context, controller) {
-              return Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text(
-                    "try_with_action".tr(),
-                    style: Theme.of(context).textTheme.displayMedium,
-                  ),
-                  Text(
-                    "tap_text_field".tr(),
-                    style: Theme.of(context).textTheme.displayMedium,
-                  ),
-                ],
-              );
-            },
-          ),
-        ],
-      ),
-    );
-
-    targets.add(
-      TargetFocus(
-        identify: "goButtonKey",
-        keyTarget: goButtonKey,
-        alignSkip: Alignment.topRight,
-        contents: [
-          TargetContent(
-            align: ContentAlign.top,
-            builder: (context, controller) {
-              return Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  Text(
-                    "step_two".tr(),
-                    style: Theme.of(context).textTheme.displayMedium,
-                  ),
-                ],
-              );
-            },
-          ),
-        ],
-      ),
-    );
-
-    return targets;
-  }
-
-  void goButtonPressed() async {
-    FirebaseAnalytics.instance.logEvent(
-      name: 'go_button_pressed',
-      parameters: <String, dynamic>{
-        "type": typeIsMovie == 0 ? "movie" : "show",
-      },
-    );
+  Future<void> _onGoPressed() async {
     FocusScope.of(context).unfocus();
-    await checkConnection();
+    await _checkConnection();
+
     if (noInternet) {
       showToastWidget(
         ToastWidget(
@@ -466,169 +229,116 @@ class _MainMenuPageState extends State<MainMenuPage> {
         ),
         duration: const Duration(seconds: 4),
       );
+      return;
+    }
+
+    if (!isLongEnough || !mounted) return;
+
+    setState(() {
+      enableLoading = true;
+      isValidQuery = false;
+    });
+
+    await _validateQuery();
+    if (!mounted) return;
+
+    if (isValidQuery) {
+      await _handleValidQuery();
     } else {
-      if (isLongEnough && mounted) {
-        setState(() {
-          enableLoading = true;
+      _handleInvalidQuery();
+    }
+  }
+
+  Future<void> _handleValidQuery() async {
+    FirebaseAnalytics.instance.logEvent(
+      name: 'valid_prompt',
+      parameters: <String, Object>{
+        "type": typeIsMovie == 0 ? "movie" : "show",
+      },
+    );
+
+    await FeedbackService.incrementSuccessfulQuery();
+
+    if (!mounted) return;
+
+    // Build the full query with settings suffix
+    final fullQuery = _controller.text + _querySettings.toPromptSuffix(isMovie: typeIsMovie == 0);
+
+    await Navigator.of(context).push(
+      MaterialPageRoute(
+        builder: (context) => RecommendationLoadingPage(
+          requestString: fullQuery,
+          type: typeIsMovie,
+          includeRentals: _querySettings.includeRentals,
+          includePurchases: _querySettings.includePurchases,
+        ),
+      ),
+    );
+
+    if (mounted) {
+      final shouldShow = await FeedbackService.shouldShowFeedbackDialog();
+      if (shouldShow && mounted) {
+        Future.delayed(const Duration(milliseconds: 300), () {
+          if (mounted) {
+            showDialog(
+              context: context,
+              barrierDismissible: false,
+              builder: (_) => const FeedbackDialog(),
+            );
+          }
         });
-        await validateQuery();
-        if (isValidQuery && mounted) {
-          FirebaseAnalytics.instance.logEvent(
-            name: 'valid_prompt',
-            parameters: <String, dynamic>{
-              "type": typeIsMovie == 0 ? "movie" : "show",
-            },
-          );
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (context) => RecommandationResultsPage(requestString: _controller.text, type: typeIsMovie),
-            ),
-          );
-        } else {
-          FirebaseAnalytics.instance.logEvent(
-            name: 'invalid_prompt',
-            parameters: <String, dynamic>{
-              "type": typeIsMovie == 0 ? "movie" : "show",
-            },
-          );
-          showToastWidget(
-            ToastWidget(
-              title: "invalid_input".tr(),
-              icon: const Icon(Icons.dangerous_outlined, color: Colors.red, size: 36),
-            ),
-            duration: const Duration(seconds: 4),
-          );
-        }
       }
     }
   }
 
-  void showExamples() {
-    showDialog(
-      context: context,
-      builder: (_) => AlertDialog(
-        shape: ShapeBorder.lerp(
-          RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-          RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-          1,
-        )!,
-        backgroundColor: Colors.grey[900]!,
-        title: Text("need_inspiration".tr()),
-        content: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Text(
-              "example_1".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_2".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_3".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_4".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_5".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-          ],
-        ),
-      ),
+  void _handleInvalidQuery() {
+    FirebaseAnalytics.instance.logEvent(
+      name: 'invalid_prompt',
+      parameters: <String, Object>{
+        "type": typeIsMovie == 0 ? "movie" : "show",
+      },
     );
-  }
 
-  void showExamplesShows() {
-    showDialog(
-      context: context,
-      builder: (_) => AlertDialog(
-        shape: ShapeBorder.lerp(
-          RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-          RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-          1,
-        )!,
-        backgroundColor: Colors.grey[900]!,
-        title: Text("need_inspiration".tr()),
-        content: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          mainAxisSize: MainAxisSize.min,
+    // Skip Firestore writes in debug mode
+    if (!kDebugMode) {
+      FirebaseFirestore.instance.collection('invalid_queries').add({
+        'type': typeIsMovie == 0 ? "movie" : "show",
+        'timestamp': FieldValue.serverTimestamp(),
+        'query': _controller.text,
+      });
+    }
+
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Row(
           children: [
-            Text(
-              "example_show_1".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
+            Expanded(
+              child: Text(
+                "invalid_query".tr(),
+                style: const TextStyle(
+                  fontSize: 15,
+                  fontWeight: FontWeight.w500,
+                  height: 1.4,
+                  color: Colors.white,
+                ),
+              ),
             ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_show_2".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_show_3".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_show_4".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
-            ),
-            const SizedBox(height: 12),
-            Container(
-              height: 1,
-              color: Colors.grey[800],
-            ),
-            const SizedBox(height: 12),
-            Text(
-              "example_show_5".tr(),
-              style: Theme.of(context).textTheme.displaySmall,
+            const SizedBox(width: 12),
+            InkWell(
+              onTap: () => ScaffoldMessenger.of(context).hideCurrentSnackBar(),
+              child: Container(
+                padding: const EdgeInsets.all(4),
+                child: const Icon(Icons.close, color: Colors.white, size: 20),
+              ),
             ),
           ],
         ),
+        backgroundColor: Colors.red,
+        behavior: SnackBarBehavior.floating,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+        margin: const EdgeInsets.all(16),
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+        duration: const Duration(seconds: 8),
       ),
     );
   }
