@@ -2,7 +2,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
 class NoInternet extends StatelessWidget {
-  const NoInternet({Key? key}) : super(key: key);
+  const NoInternet({super.key});
 
   @override
   Widget build(BuildContext context) {
