@@ -1,17 +1,17 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:http/http.dart' as http;
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:watch_next/pages/main_menu_page.dart';
+import 'package:watch_next/pages/home_page.dart';
 import 'package:watch_next/services/database_service.dart';
 import 'package:watch_next/services/http_service.dart';
+import 'package:watch_next/services/query_cache_service.dart';
 
 ///This is the page where you enter the movie you saw
 class StreamingServicesPage extends StatefulWidget {
   const StreamingServicesPage({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
   @override
   State<StreamingServicesPage> createState() => _StreamingServicesPage();
 }
@@ -23,7 +23,7 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
 
   @override
   void initState() {
-    resultList = HttpService().getWatchProvidersByLocale(http.Client());
+    resultList = HttpService().getWatchProvidersByLocale();
     DatabaseService.getAllStreamingServices().then(
       (mapList) => {
         for (var map in mapList)
@@ -41,7 +41,7 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: const Color.fromRGBO(11, 14, 23, 1),
+      backgroundColor: Theme.of(context).colorScheme.primary,
       body: body(),
     );
   }
@@ -49,17 +49,24 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
   Widget body() {
     return Column(
       children: [
-        const SizedBox(height: 48),
-        Text(
-          "select_streaming".tr(),
-          textAlign: TextAlign.center,
-          style: Theme.of(context).textTheme.displayMedium,
+        const SizedBox(height: 60),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 24),
+          child: Column(
+            children: [
+              Text(
+                "select_streaming".tr(),
+                textAlign: TextAlign.center,
+                style: Theme.of(context).textTheme.displayLarge,
+              ),
+            ],
+          ),
         ),
         const SizedBox(height: 32),
-        streamingGrid(),
         Expanded(
-          child: Container(),
+          child: streamingGrid(),
         ),
+        const SizedBox(height: 16),
         closeButton(),
         const SizedBox(height: 32),
       ],
@@ -70,34 +77,80 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
     return FutureBuilder<dynamic>(
       future: resultList,
       builder: (context, snapshot) {
+        if (snapshot.hasError) {
+          return Center(
+            child: Padding(
+              padding: const EdgeInsets.all(24),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Container(
+                    width: 80,
+                    height: 80,
+                    decoration: BoxDecoration(
+                      color: Colors.red.withValues(alpha: 0.1),
+                      shape: BoxShape.circle,
+                    ),
+                    child: Icon(Icons.error_outline, size: 48, color: Colors.red),
+                  ),
+                  SizedBox(height: 24),
+                  Text(
+                    "error_occurred".tr(),
+                    style: Theme.of(context).textTheme.displayMedium,
+                    textAlign: TextAlign.center,
+                  ),
+                  SizedBox(height: 16),
+                  ElevatedButton(
+                    onPressed: () {
+                      setState(() {
+                        resultList = HttpService().getWatchProvidersByLocale();
+                      });
+                    },
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor: Colors.orange,
+                      foregroundColor: Colors.white,
+                      padding: EdgeInsets.symmetric(horizontal: 32, vertical: 16),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(12),
+                      ),
+                    ),
+                    child: Text(
+                      "Try Again",
+                      style: TextStyle(fontWeight: FontWeight.bold),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          );
+        }
+
         if (snapshot.hasData && snapshot.data.length > 0) {
           return Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 12),
+            padding: const EdgeInsets.symmetric(horizontal: 24),
             child: Container(
               decoration: BoxDecoration(
-                color: Colors.grey[900],
+                color: Theme.of(context).colorScheme.tertiary,
+                borderRadius: BorderRadius.circular(16),
                 border: Border.all(
-                  color: Colors.grey[700]!,
+                  color: Theme.of(context).colorScheme.outline,
+                  width: 1,
                 ),
-                borderRadius: BorderRadius.circular(8),
               ),
-              padding: const EdgeInsets.only(top: 12, left: 12, right: 12),
-              height: MediaQuery.of(context).size.height * 0.7,
-              child: GridView.builder(
-                padding: EdgeInsets.zero,
-                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-                  crossAxisCount: 3,
-                  childAspectRatio: 1,
-                ),
-                itemCount: snapshot.data.length,
-                itemBuilder: (BuildContext context, int index) {
-                  return Padding(
-                    padding: const EdgeInsets.all(8),
-                    child: TextButton(
-                      style: TextButton.styleFrom(
-                        padding: const EdgeInsets.all(0),
-                      ),
-                      onPressed: () {
+              child: ClipRRect(
+                borderRadius: BorderRadius.circular(16),
+                child: GridView.builder(
+                  padding: const EdgeInsets.all(16),
+                  gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                    crossAxisCount: 3,
+                    childAspectRatio: 1,
+                    crossAxisSpacing: 12,
+                    mainAxisSpacing: 12,
+                  ),
+                  itemCount: snapshot.data.length,
+                  itemBuilder: (BuildContext context, int index) {
+                    return InkWell(
+                      onTap: () {
                         setState(() {
                           if (selectedStreamingServices.containsKey(snapshot.data[index].providerId)) {
                             selectedStreamingServices
@@ -107,16 +160,19 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
                           }
                         });
                       },
+                      borderRadius: BorderRadius.circular(16),
                       child: gridItem(snapshot.data[index].logoPath, index, snapshot.data[index].providerId),
-                    ),
-                  );
-                },
+                    );
+                  },
+                ),
               ),
             ),
           );
         } else {
-          return Expanded(
-            child: Container(),
+          return Center(
+            child: CircularProgressIndicator(
+              color: Colors.orange,
+            ),
           );
         }
       },
@@ -125,28 +181,54 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
 
   Widget closeButton() {
     return selectedStreamingServices.isNotEmpty
-        ? SizedBox(
-            height: 50,
-            child: Center(
-              child: TextButton(
-                onPressed: () async {
+        ? Container(
+            margin: const EdgeInsets.symmetric(horizontal: 24),
+            width: double.infinity,
+            height: 56,
+            decoration: BoxDecoration(
+              gradient: LinearGradient(
+                colors: [Colors.orange, Colors.orange[700]!],
+                begin: Alignment.topLeft,
+                end: Alignment.bottomRight,
+              ),
+              borderRadius: BorderRadius.circular(16),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.orange.withValues(alpha: 0.3),
+                  blurRadius: 12,
+                  offset: const Offset(0, 4),
+                ),
+              ],
+            ),
+            child: Material(
+              color: Colors.transparent,
+              child: InkWell(
+                borderRadius: BorderRadius.circular(16),
+                onTap: () async {
                   SharedPreferences prefs = await SharedPreferences.getInstance();
                   bool seen = prefs.getBool('skip_intro') ?? false;
                   prefs.setBool('skip_intro', true);
                   await DatabaseService.saveStreamingServices(selectedStreamingServices);
-                  if (context.mounted && seen) {
+                  // Clear query cache since streaming services changed
+                  await QueryCacheService.clearAllCaches();
+                  if (mounted && seen) {
                     Navigator.of(context).pop();
                   } else if (mounted && !seen) {
                     Navigator.of(context).pushReplacement(
                       MaterialPageRoute(
-                        builder: (context) => const MainMenuPage(),
+                        builder: (context) => const TabNavigationPage(),
                       ),
                     );
                   }
                 },
-                child: Text(
-                  "done".tr(),
-                  style: Theme.of(context).textTheme.bodyMedium,
+                child: Center(
+                  child: Text(
+                    "done".tr().toUpperCase(),
+                    style: Theme.of(context).textTheme.labelMedium?.copyWith(
+                          fontWeight: FontWeight.bold,
+                          letterSpacing: 1.2,
+                        ),
+                  ),
                 ),
               ),
             ),
@@ -155,33 +237,79 @@ class _StreamingServicesPage extends State<StreamingServicesPage> with TickerPro
   }
 
   Widget gridItem(String logo, int index, int providerId) {
-    return Container(
-      height: 100,
-      width: 100,
+    bool isSelected = selectedStreamingServices.keys.contains(providerId);
+
+    return AnimatedContainer(
+      duration: const Duration(milliseconds: 200),
       decoration: BoxDecoration(
-        borderRadius: const BorderRadius.all(Radius.circular(20)),
+        borderRadius: BorderRadius.circular(16),
         border: Border.all(
           width: 3,
-          color: selectedStreamingServices.keys.contains(providerId) ? Colors.orange : Colors.white,
+          color: isSelected ? Colors.orange : Colors.grey[700]!,
         ),
-        color: Colors.grey[300],
+        color: Theme.of(context).colorScheme.tertiary,
+        boxShadow: isSelected
+            ? [
+                BoxShadow(
+                  color: Colors.orange.withValues(alpha: 0.3),
+                  blurRadius: 8,
+                  offset: const Offset(0, 2),
+                ),
+              ]
+            : [],
       ),
-      child: Center(
-        child: ClipRRect(
-          borderRadius: BorderRadius.circular(16),
-          child: CachedNetworkImage(
-            fit: BoxFit.cover,
-            imageUrl: "https://image.tmdb.org/t/p/original//$logo",
-            placeholder: (context, url) => Container(
-              color: const Color.fromRGBO(11, 14, 23, 1),
-            ),
-            errorWidget: (context, url, error) => Expanded(
-              child: Container(
-                color: Colors.grey[800],
+      child: Stack(
+        children: [
+          Center(
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(12),
+              child: CachedNetworkImage(
+                fit: BoxFit.cover,
+                imageUrl: "https://image.tmdb.org/t/p/original//$logo",
+                placeholder: (context, url) => Container(
+                  color: Theme.of(context).colorScheme.outline,
+                  child: Center(
+                    child: CircularProgressIndicator(
+                      strokeWidth: 2,
+                      color: Colors.grey[600],
+                    ),
+                  ),
+                ),
+                errorWidget: (context, url, error) => Container(
+                  color: Theme.of(context).colorScheme.tertiary,
+                  child: Icon(
+                    Icons.broken_image,
+                    color: Colors.grey[600],
+                  ),
+                ),
               ),
             ),
           ),
-        ),
+          if (isSelected)
+            Positioned(
+              top: 4,
+              right: 4,
+              child: Container(
+                width: 24,
+                height: 24,
+                decoration: BoxDecoration(
+                  color: Colors.orange,
+                  shape: BoxShape.circle,
+                  boxShadow: [
+                    BoxShadow(
+                      color: Colors.black.withValues(alpha: 0.3),
+                      blurRadius: 4,
+                    ),
+                  ],
+                ),
+                child: const Icon(
+                  Icons.check_rounded,
+                  size: 16,
+                  color: Colors.white,
+                ),
+              ),
+            ),
+        ],
       ),
     );
   }
